factor out ability triggers, remove redundant scale functions to use Unity transform
