+++ Assets/Prefabs/Ability/Abilities/AbilityObstacle.prefab
@@ -14,7 +14,6 @@ GameObject:
   - component: {fileID: 901539224883860523}
   - component: {fileID: 75311920755393853}
   - component: {fileID: -6358432458901213615}
-  - component: {fileID: -4243902514752060037}
   m_Layer: 6
   m_Name: AbilityObstacle
   m_TagString: Untagged
@@ -32,7 +31,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalScale: {x: 3, y: 3, z: 3}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
@@ -146,16 +145,3 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   timeBeforeDespawn: 5
---- !u!114 &-4243902514752060037
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4873085143038728622}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 48caed4633b3a9646becab2844747340, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  dimensions: {x: 3, y: 3, z: 3}
