+++ Assets/Scripts/Ability/AbMoveTowardsDst.cs
@@ -2,14 +2,13 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[RequireComponent(typeof(Ability), typeof(Collider), typeof(Rigidbody))]
-public class AbilityMoveTowardsDst : MonoBehaviour
+[RequireComponent(typeof(Ability))]
+public class AbMoveTowardsDst : MonoBehaviour
 {
     [SerializeField] float moveSpeed = 15f;
     [SerializeField] bool isHoming = false;
 
-    private bool _hasOriented = false;
-
+    bool _hasOriented = false;
     Ability _ability;
 
     private void Start()
@@ -32,13 +31,13 @@ public class AbilityMoveTowardsDst : MonoBehaviour
     {
         if (!_hasOriented || isHoming)
         {
-            transform.LookAt(_ability.DstPos);
+            _ability.transform.LookAt(_ability.DstPos);
             _hasOriented = true;
         }
     }
 
     void Move()
     {
-        transform.position = transform.position + moveSpeed * Time.deltaTime * transform.forward;
+        _ability.transform.position = _ability.transform.position + moveSpeed * Time.deltaTime * _ability.transform.forward;
     }
 }
