+++ Assets/Scripts/Ability/AbPositionAtSrcTowardsDst.cs
@@ -3,14 +3,15 @@ using System.Collections.Generic;
 using UnityEngine;
 
 [RequireComponent(typeof(Ability))]
-public class AbilityPositionInFrontOfSrcTowardsDst : MonoBehaviour
+public class AbPositionAtSrcTowardsDst : MonoBehaviour
 {
     Ability _ability;
 
     private void Start()
     {
         _ability = GetComponent<Ability>();
+
+        transform.position = _ability.SrcPos;
         transform.LookAt(_ability.DstPos);
-        transform.position = _ability.SrcPos + transform.localScale.z / 2 * transform.forward;
     }
 }
