--- Assets/Scripts/Ability/AbilityScaleToBox.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 48caed4633b3a9646becab2844747340
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
