--- Assets/Scripts/Ability/AbilityScaleToCylinder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 834175a1cef6ea54fa190aaf7347c212
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
