--- Assets/Scripts/Ability/AbilityScaleToSphere.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a574d6c986b15084185d8edda7079eb7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
